fix(output): fix crash with ogg files
When a new ogg file would start to play, the frame count per packet
would naturally drop due to the buffered decoder. However, since the
current index is not updated in the loop, this change wouldn't be shown.
The fix is to return the current index of the player alongside the
current packet.